added modal to delete button and made room for it in article tag

+++ src/main/java/com/example/demo/controller/LoginController.java
@@ -133,9 +133,19 @@ public class LoginController {
     }
     //Delete Plant
     @GetMapping("/delete/{id}")
-    public String delete(@PathVariable("id") String id) throws Exception {
-        plantService.deletePlant(id);
-        return "redirect:/home";
+    public String delete(@PathVariable("id") Long id,RedirectAttributes ra) throws Exception {
+        Admin admin = getLoggedInAdmin();
+        List<Plant> userPlants = plantRepository.findAllByAdminId(admin.getId());
+        Plant plant = plantRepository.findById(id).get();
+        if (userPlants.contains(plant)) {
+            plantService.deletePlant(id);
+            ra.addFlashAttribute("SuccesPlantCreation", "Your plant has been removed");
+            return "redirect:/home";
+        } else {
+            ra.addFlashAttribute("ErrorPlantRemoval", "No such plant, Master");
+            return "redirect:/home";
+        }
+
     }
 
 }
